Use float64 precision during truncation to avoid quantization loss
Upconvert factors to float64 at the start of truncation to eliminate
quantization errors during orthogonalization, SVD, and scaling operations.
The constructor automatically downconverts back to the original dtype
(typically bfloat16 or float32) when creating the final TensorizedLinear.

This ensures truncation operates at maximum numerical precision.

+++ src/llmcompressor/modifiers/experimental/tensorized_linear.py
@@ -347,7 +347,8 @@ class TensorizedLinear(nn.Module):
             New TensorizedLinear with reduced ranks
         """
         orig_device = self.factors[0].device
-        factors = [f.detach().to(torch.float32) for f in self.factors]
+        # Upconvert to float64 for high-precision truncation to avoid quantization loss
+        factors = [f.detach().to(torch.float64) for f in self.factors]
         num_cores = len(factors)
 
         # Compute Frobenius norm of original matrix before truncation
